requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.21.0
opencv-python>=4.5.0
mediapipe>=0.8.0
google-genai>=0.3.0
sentence-transformers>=2.2.0
python-dotenv>=0.19.0
keyboard>=0.13.5
scikit-learn>=1.0.0
matplotlib>=3.5.0
tqdm>=4.62.0
seaborn>=0.11.0
